Update smart-contract guide instruction. (#94)
This instruction: ”_Open your project in a code editor and delete the
boilerplate code in src/main.sw so that you start with an empty file_.”
was caussing the language server to crash with the following error:

```
thread 'main' panicked at 'called `Option::unwrap()` on a `None` value', 
sway-parse/src/parser.rs:33:18
```

It seems that the parser in the compiler panics on an empty file which
is causing the language server to crash. I tested this on master and the
problem has since been resolved.

This PR updates the instruction to delete everything apart from the
first line in order for the language server to not crash and be
available for the rest of the tutorial.